pyproject.toml
[project]
name = "anim-ml"
version = "0.1.0"
description = "ML model training and gRPC server for animation curve suggestion system"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.3",
    "numpy>=1.26",
    "onnx>=1.16",
    "onnxruntime>=1.18",
    "h5py>=3.11",
    "scipy>=1.13",
    "pyyaml>=6.0",
    "onnxscript>=0.6.2",
]

[project.optional-dependencies]
train = [
    "mlflow>=2.14",
    "matplotlib>=3.9",
    "tensorboard>=2.17",
]
server = [
    "grpcio>=1.64",
    "grpcio-tools>=1.64",
    "protobuf>=5.27",
]
dev = [
    "pytest>=8.2",
    "ruff>=0.5",
    "pyright>=1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/anim_ml"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "TCH"]

[tool.ruff.lint.per-file-ignores]
"src/anim_ml/server/proto/animation_ml_pb2.py" = ["E", "F", "W", "I", "UP", "B", "SIM", "TCH"]
"src/anim_ml/server/proto/animation_ml_pb2_grpc.py" = ["E", "F", "W", "I", "UP", "B", "SIM", "TCH"]
"src/anim_ml/server/proto/animation_ml_pb2.pyi" = ["E", "F", "W", "I", "UP", "B", "SIM", "TCH"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
exclude = [
    "src/anim_ml/server/proto/animation_ml_pb2.py",
    "src/anim_ml/server/proto/animation_ml_pb2_grpc.py",
    "src/anim_ml/server/proto/animation_ml_pb2.pyi",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Fast unit tests, no external data needed",
    "integration: Requires downloaded BVH data or HDF5 files",
    "slow: Full training or large dataset processing",
    "gpu: Requires CUDA GPU",
    "docker: Requires Docker daemon",
]
